requirements.txt
Flask==2.2.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-Testing==0.8.1
python-decouple==3.5
mysql-connector-python==8.1.0
SQLAlchemy==2.0.21
requests==2.31.0
Werkzeug==2.2.0
